cgimap: Normalize env variables for DB. (#18)
Doing this cleans things opens up the potential to clean things
up in several ways, for example having one consolidated
environment file.

We can also now share a postgres-wait script in between the three
containers (postgis, cgimap, rails) that need to wait for the
database to be ready and open for queries before initializing.