rails/cgimap/osmosis: Postgres wait consolidation (#18)
It all feels so much cleaner now.

I tried to use a symlink to the shared scripts directory, but
Docker doesn't roll that way. ;_;

I wish we could use this technique to share the Rails schema
with the Postgis container, but that would create a circular
dependency with the Rails container.